refactor(files): split document extraction into provider-attachment-extraction.ts (size budget)
